plugins/homeassistant/pyproject.toml
[project]
name = "coremind-plugin-homeassistant"
version = "0.1.0"
description = "Home Assistant integration plugin for CoreMind."
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
authors = [{ name = "Guillaume Gagnon" }]
requires-python = ">=3.12"
keywords = ["coremind", "plugin", "home-assistant"]

dependencies = [
    "aiohttp>=3.10.0",
    "coremind>=0.0.1",
    "grpcio>=1.66.0",
    "structlog>=24.4.0",
]

[project.scripts]
coremind-plugin-homeassistant = "coremind_plugin_homeassistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coremind_plugin_homeassistant"]
